+++ libs/vault/src/cipher-view/card-details/card-details-view.component.html
@@ -1,6 +1,3 @@
-<bit-callout *ngIf="cardIsExpired" type="info" [title]="'cardExpiredTitle' | i18n">
-  {{ "cardExpiredMessage" | i18n }}
-</bit-callout>
 <bit-section>
   <bit-section-header>
     <h2 bitTypography="h6">{{ setSectionTitle }}</h2>
