+++ libs/vault/src/cipher-view/cipher-view.component.ts
@@ -8,10 +8,11 @@ import { Organization } from "@bitwarden/common/admin-console/models/domain/orga
 import { CollectionId } from "@bitwarden/common/types/guid";
 import { CollectionService } from "@bitwarden/common/vault/abstractions/collection.service";
 import { FolderService } from "@bitwarden/common/vault/abstractions/folder/folder.service.abstraction";
+import { CardView } from "@bitwarden/common/vault/models/view/card.view";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 import { CollectionView } from "@bitwarden/common/vault/models/view/collection.view";
 import { FolderView } from "@bitwarden/common/vault/models/view/folder.view";
-import { SearchModule } from "@bitwarden/components";
+import { SearchModule, CalloutModule } from "@bitwarden/components";
 
 import { AdditionalOptionsComponent } from "./additional-options/additional-options.component";
 import { AttachmentsV2ViewComponent } from "./attachments/attachments-v2-view.component";
@@ -28,6 +29,7 @@ import { ViewIdentitySectionsComponent } from "./view-identity-sections/view-ide
   templateUrl: "cipher-view.component.html",
   standalone: true,
   imports: [
+    CalloutModule,
     CommonModule,
     SearchModule,
     JslibModule,
@@ -48,6 +50,7 @@ export class CipherViewComponent implements OnInit, OnDestroy {
   folder$: Observable<FolderView>;
   collections$: Observable<CollectionView[]>;
   private destroyed$: Subject<void> = new Subject();
+  cardIsExpired: boolean = false;
 
   constructor(
     private organizationService: OrganizationService,
@@ -57,6 +60,8 @@ export class CipherViewComponent implements OnInit, OnDestroy {
 
   async ngOnInit() {
     await this.loadCipherData();
+
+    this.cardIsExpired = this.isCardExpiryInThePast();
   }
 
   ngOnDestroy(): void {
@@ -96,5 +101,25 @@ export class CipherViewComponent implements OnInit, OnDestroy {
         .getDecrypted$(this.cipher.folderId)
         .pipe(takeUntil(this.destroyed$));
     }
+  }
+
+  isCardExpiryInThePast() {
+    if (this.cipher.card) {
+      const { expMonth, expYear }: CardView = this.cipher.card;
+
+      if (expYear && expMonth) {
+        // `Date` months are zero-indexed
+        const parsedMonth = parseInt(expMonth) - 1;
+        const parsedYear = parseInt(expYear);
+
+        // First day of the next month minus one, to get last day of the card month
+        const cardExpiry = new Date(parsedYear, parsedMonth + 1, 0);
+        const now = new Date();
+
+        return cardExpiry < now;
+      }
+    }
+
+    return false;
   }
 }
